getCountdown
