+++ packages/metavinci/src/components/ItemCard/index.tsx
@@ -5,6 +5,7 @@ import { Card, Row, Col } from 'antd'
 import { Art } from '../../types'
 
 import './index.less'
+import { getCountdown } from '../../utils/utils'
 
 
 export const ItemCard = ({ art }: { art: Art }) => {
@@ -16,19 +17,7 @@ export const ItemCard = ({ art }: { art: Art }) => {
   useEffect(() => {
     if (art.endingTS) {
       const interval = setInterval(() => {
-        const now = (new Date()).getTime()
-        let delta = Math.abs(art.endingTS as number - now) / 1000
-  
-        const days = Math.floor(delta / 86400)
-        delta -= days * 86400
-  
-        const hours = Math.floor(delta / 3600) % 24
-        delta -= hours * 3600
-  
-        const minutes = Math.floor(delta / 60) % 60
-        delta -= minutes * 60
-
-        const seconds = Math.floor(delta % 60)
+        const {minutes, seconds} = getCountdown(art.endingTS as number)
   
         setMinutes(minutes)
         setSeconds(seconds)
